Remove unused abstractions in spatial doc values
